feat(diarias-turma): permite valores quebrados em Quantidade de Pessoas
- prisma/schema.prisma: DiariaTurma.quantidadePessoas Int -> Float
  (banco de produção já alterado manualmente via ALTER TABLE, confirmado
  pelo usuário antes deste commit)
- DiariaTurmaForm.tsx: parseInt -> parseFloat no cálculo de valorTotal e
  no payload enviado; step do input 1 -> 0.5 (permite incrementos de
  meio período, ex: 8, 8.5, 9)
- app/api/diarias-turma/route.ts e [id]/route.ts: parseInt -> parseFloat
  ao ler quantidadePessoas do body
- app/modules/turmas/page.tsx: exibição condicional na tabela — mostra
  inteiro sem casas decimais (ex: '9') e só usa toFixed(1) quando o
  valor é quebrado (ex: '8.5')

Nota: não foi possível rodar 'npx prisma generate' neste sandbox
(binaries.prisma.sh bloqueado pela allowlist de rede do ambiente) —
mesma limitação já vista antes. O client Prisma será gerado
normalmente no build da Vercel, que tem acesso irrestrito à rede.

+++ app/api/diarias-turma/[id]/route.ts
@@ -55,7 +55,7 @@ export async function PUT(
 
       const body = await request.json()
 
-      const quantidadePessoas = body.quantidadePessoas !== undefined ? parseInt(body.quantidadePessoas) : diaria.quantidadePessoas
+      const quantidadePessoas = body.quantidadePessoas !== undefined ? parseFloat(body.quantidadePessoas) : diaria.quantidadePessoas
           const valorDiaria = body.valorDiaria !== undefined ? parseFloat(body.valorDiaria) : diaria.valorDiaria
           const valorTotal = parseFloat((valorDiaria * quantidadePessoas).toFixed(2))
